Reproduction for MockedProvider issue
In src/App.spec.js I've provided two reproductions.

The first is following the steps from
https://www.apollographql.com/docs/react/v3.0-beta/development-testing/testing/
as closely as possible with an illustration of the order dependence of
the issue.

The second is a minimal reproduction.

You are safe to ignore the warnings the problem is still present with
all of them resolved. It just takes more code and would obscure the
example unnecessarily.

The issue is also present in @apollo/react-testing@3.1.4 with apollo-client@2.6.10
as well as if you switch out react-test-renderer for enzyme.